+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 Nothing yet.
 
+## 1.1.0
+
+Released Sun 3rd Jan 2016.
+
+### New
+
+* Added support for wrapping Stringy data.
+  * Added `GanbaroDigital\Guarantees\Values\StringGuarantee`
+
 ## 1.0.0
 
 Released Sat 2nd Jan 2016.